Improved scripts

+++ Classes/CodeToHtml.sh
@@ -1,10 +1,18 @@
 #!/bin/bash
+CODE_TO_HTML_BIN=~/bin/ToolCodeToHtmlConsole
+
+if [ ! -e $CODE_TO_HTML_BIN ]
+then
+  echo "ERROR: CodeToHtml binary not found at "$CODE_TO_HTML_BIN
+  exit
+fi
+
 for folder in `ls | egrep "Cpp"`
 do
   echo $folder
-
-  ~/bin/ToolCodeToHtmlConsole ./$folder
+  $CODE_TO_HTML_BIN ./$folder
 done
 
 rm tmp.txt
 rm trace_out.txt
+mv *.htm ~/GitHubs/RichelbilderbeekNl/Cpp
